Refactor the helpers into separate files.
