+++ containers/flask/app/puzzles.py
@@ -1,7 +1,7 @@
 """This module gets puzzles out of the database."""
 from flask_pymongo import PyMongo
 from app import app
-from .helpers import safe_list_get
+from app.helpers.formatters import safe_list_get
 
 app.config["MONGO_URI"] = "mongodb://mongo/codeselfstudy"
 mongo = PyMongo(app)
