+++ containers/flask/app/slack/puzzle_command.py
@@ -6,7 +6,7 @@ import hmac
 import time
 import hashlib
 from os import environ
-from app.helpers import safe_list_get
+from app.helpers.utilities import safe_list_get
 from urllib.parse import parse_qs
 
 # Example payload:
